+++ src/client/canvas.ts
@@ -5,27 +5,25 @@ export const penCanvas = document.createElement('canvas');
 export const penCtx = penCanvas.getContext('2d')!;
 penCanvas.id = 'pen-canvas';
 
-let ratio: number = 16 / 9;
-let scale: number;
+export let aspectRatio: number = 16 / 9;
+export let scale: number;
 
-export function setScale(newScale: number) {
-    scale = newScale;
-
-    canvas.width = ratio * scale;
+function updateSize() {
+    canvas.width = aspectRatio * scale;
     canvas.height = scale;
 
-    penCanvas.width = ratio * scale;
+    penCanvas.width = aspectRatio * scale;
     penCanvas.height = scale;
 }
 
+export function setScale(newScale: number) {
+    scale = newScale;
+    updateSize();
+}
+
 export function setAspectRatio(newAspectRatio: number) {
-    ratio = newAspectRatio;
-
-    canvas.width = ratio * scale;
-    canvas.height = scale;
-
-    penCanvas.width = ratio * scale;
-    penCanvas.height = scale;
+    aspectRatio = newAspectRatio;
+    updateSize();
 }
 
 canvas.parentElement?.insertBefore(penCanvas, canvas);