+++ tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'tsup';
 
 export default defineConfig({
-    entry: ['src/index.ts'],
+    entry: ['src/client/index.ts', 'src/server/index.ts'],
     format: ['esm', 'cjs'],
     dts: true,
     sourcemap: true,